Move and rename elements ArrowPopover
